# microservices/employee-management/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Employee Management API
  version: "1.0"
  description: API for managing employees
  termsOfService: http://swagger.io/terms/
  contact:
    name: API Support
    email: josed.amayar@uqvirtual.edu.co

servers:
  - url: /employees-service/api

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    api.ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        rejectedValue: {}
    api.ErrorResponse:
      type: object
      description: Standard error response structure
      properties:
        error:
          type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/api.ErrorDetail'
        message:
          type: string
        path:
          type: string
        status:
          type: integer
        timestamp:
          type: string
    api.PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/models.Employee'
        pagination:
          $ref: '#/components/schemas/api.PaginationMeta'
    api.PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
        total_records:
          type: integer
    models.Employee:
      type: object
      properties:
        createdAt:
          type: string
        departmentID:
          type: string
        email:
          type: string
        hireDate:
          type: string
        id:
          type: integer
        name:
          type: string
        status:
          $ref: '#/components/schemas/models.EmployeeStatus'
        updatedAt:
          type: string
        role:
          type: string
    models.EmployeeStatus:
      type: string
      enum:
        - ACTIVE
        - ON_VACATION
        - RETIRED

security:
  - BearerAuth: []

paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
      security: [] # No auth required for health check
  /employees:
    get:
      summary: Get all employees with pagination and filtering
      tags:
        - Employees
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          description: Page number (default 1)
        - name: page_size
          in: query
          schema:
            type: integer
          description: Number of items per page
        - name: department
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.PaginatedResponse'
        "401":
          description: Unauthorized
    post:
      summary: Create a new employee
      tags:
        - Employees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                departmentID:
                  type: string
                role:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Employee'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (ADMIN required)
  /employees/{id}:
    get:
      summary: Get employee by ID
      tags:
        - Employees
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Employee'
    put:
      summary: Update employee
      tags:
        - Employees
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.Employee'
      responses:
        "200":
          description: OK
    delete:
      summary: Delete employee
      tags:
        - Employees
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content